creates campsites route and moves enpoints
